+++ src/Session/AnimationObject.h
