+++ src/Session/SessionManager.h
